This is my first online website!
This is my first project that has been realized in an online environment. It is a site that should only display the products that my client offers. The site has almost no functionality it is more informative and its purpose is to refer people to my client's Instagram. It is made for Bulgarian users only.